+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 2.0.0 (2020-08-25)
 
 * **[BC]** Bump minimum PHP version to 7.2
-* **[BC]** Remove `Icecave\Chrono\PackageInfo` class
+* **[BC]** Remove `PackageInfo` class
 
 ### 1.0.4 (2016-02-19)